% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.r
\name{combine.morpho}
\alias{combine.morpho}
\title{Combine Two Morpho Objects}
\usage{
combine.morpho(x, y)
}
\arguments{
\item{x}{A `morpho` object.}

\item{y}{A `morpho` object.}
}
\value{
A combined `morpho` object.
}
\description{
This function merges two `morpho` objects, combining their sequences,
model parameters, and transition histories, while ensuring tree and fossil
consistency.
}
\examples{
# combined <- combine_morpho(morpho1, morpho2)
}
