% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.r
\name{convergent_evol}
\alias{convergent_evol}
\title{Determines the number of convergently evolved traits}
\usage{
convergent_evol(data = NULL)
}
\arguments{
\item{data}{a morpho object

 @export}
}
\description{
This function determines which traits have evoled through convergent evolution
}
