% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degrade.data.R
\name{degrade.data}
\alias{degrade.data}
\title{Degrade morphological data}
\usage{
degrade.data(data = NULL, method = NULL, percentage = NULL)
}
\arguments{
\item{data}{A morpho object with sequence data}

\item{method}{This specifies under what criteria data should be removed. Methods include "random" which randomly removes
characters across the entire matrix, "partition" which allows different partitions (that were used to simualte the data)
to loose varying amounts of characters, "rates" which removes characters depening on the rate the were simulated under,
and "character" where the user can specify particular characters to degrade the data.}

\item{percentage}{This specifies how much data to remove. The number of percentages supplied varies depending on the
method chosen. For random the user must supply 1 percentage, and for all other methods must supply 1 percentage
per category.}
}
\value{
An object of class morpho.
}
\description{
This function removes characters from a morphological matrix simulated using morphosim
}
\examples{
# simulated tree
phy <- ape::rtree(10)

# simulate characters along the branches of the tree
transition_history <-  sim.morpho.history(tree = phy,
                                         k = 3,
                                         trait.num = 30,
                                         ancestral = TRUE,
                                         ACRV = "gamma",
                                         variable = FALSE)

# randomly remove data
degraded <- degrade.data(data = transition_history, method = "random", percentage = 10)


# remove data based on the rate it was simulated under
degraded <- degrade.data(data = transition_history, method = "rate", percentage = c(0,0,20,100))
}
