% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.r
\name{find_path_to_tip}
\alias{find_path_to_tip}
\title{Determines the route (nodes and branches) for a tip in a phylogenetic tree}
\usage{
find_path_to_tip(tree, tip)
}
\arguments{
\item{tree}{phylogenetic tree}

\item{tip}{tip label}
}
\description{
This function traverses the tree to determine the evolutionary path (branches)
from root to a given tip
}
\examples{
route_n <- find_path_to_tip(tree, "t2")

}
