% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.r
\name{get.reconstructed}
\alias{get.reconstructed}
\title{Add reconstructed tree and matrix to morpho object}
\usage{
get.reconstructed(data)
}
\arguments{
\item{data}{`morpho object` containing a fossil object}
}
\value{
a `morpho object`
}
\description{
Function to add the reconstructed tree and corresponding reconstructed
matrix to an existing morpho object
}
\examples{
#' # simulate a phylogenetic tree
t = TreeSim::sim.bd.taxa(n = 5, numbsim = 1, lambda = 0.1, mu = 0.05)[[1]]

# simulate fossil sampling
f = FossilSim::sim.fossils.poisson(rate = 0.1, tree = t, root.edge = F)

# simulate characters along the branches of the tree
morpho_data <-  sim.morpho(tree = phy,
                           k = c(2,3,4),
                           trait.num = 20,
                           ancestral = TRUE,
                           partition = c(10,5,5),
                           ACRV = "gamma",
                           variable = TRUE,
                           ACRV.ncats = 4,
                           define.Q = NULL,
                           fossil = f)

re <- get_reconstructed(morpho_data)

}
