% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.colours.R
\name{get_colours}
\alias{get_colours}
\title{#' @description
This function is used to colour the branches in the plotting function when plotting the reconstructed tree.}
\usage{
get_colours(data)
}
\description{
#' @description
This function is used to colour the branches in the plotting function when plotting the reconstructed tree.
}
