% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morpho.R
\name{morpho}
\alias{morpho}
\alias{as.morpho}
\alias{is.morpho}
\title{Mopho object}
\usage{
morpho(data = NULL, tree = NULL, time.tree = NULL, model = NULL)

as.morpho(data, tree, model)

is.morpho(data)
}
\arguments{
\item{data}{?}

\item{tree}{Tree with branches that represent genetic distance associated with the character data.}

\item{time.tree}{Tree with branches that represent time associated with the character data.}
}
\description{
Create a morpho.
}
