% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.r
\name{morphsim_fossilsim}
\alias{morphsim_fossilsim}
\title{Match sampled ancestor labels}
\usage{
morphsim_fossilsim(data = NULL)
}
\arguments{
\item{data}{Morpho object containing fossils}
}
\description{
Match the sampled ancestor labels from Morphsim and Fossilsim
}
\examples{
morphsim_fossilsim <- function(data = morpho_data)

}
