% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.morpho.grid.R
\name{plot.morpho.grid}
\alias{plot.morpho.grid}
\title{This function plots your morpho object to a character matrix}
\usage{
\method{plot}{morpho.grid}(
  x,
  xlab = "",
  ylab = "",
  name = "",
  col = c("white", "gray", "lightblue", "pink", "yellow", "green", "orange")
)
}
\arguments{
\item{x}{an object of class "morpho"}

\item{col}{a string bound by c() containing colors to match the amount of possible character states}
}
\description{
This function plots your morpho object to a character matrix
}
