% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.morpho.grid.R
\name{plotMorphoGrid}
\alias{plotMorphoGrid}
\title{Plots morphological matrix}
\usage{
plotMorphoGrid(
  data = NULL,
  num.trait = "all",
  col = c("#fdfdfd", "lightgray", "lightblue", "pink", "yellow", "green", "orange")
)
}
\arguments{
\item{data}{A morpho object}

\item{num.trait}{default is set to "all" which plots all traits in black font. If you
want to focus on a specific trait set it here, e.g. num.trait = 1 and this trait will
be highlighted}
}
\description{
This function plots the full morphological matrix assocaited with the character data
at the tips of a tree. Requires a moprho object as input.
}
\examples{
 # simualte a tree
phy <- ape::rtree(10)

# simulate character data for the tree
simulated_morpho <- sim.morpho(phy, k = 2, trait.num = 10)

# plot the character matrix
plotMorphoGrid(data = simulated_morpho, num.trait = 1)

}
