% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.morpho.completeprocess.R
\name{sim.morpho.completeprocess}
\alias{sim.morpho.completeprocess}
\title{Simulate data along branches}
\usage{
sim.morpho.completeprocess(
  tree = NULL,
  time.tree = NULL,
  br.rates = NULL,
  k = 2,
  trait.num = 2
)
}
\arguments{
\item{tree}{Tree with branches that represent genetic distance associated with the character data.}

\item{time.tree}{Tree with branches that represent time associated with the character data.}

\item{k}{Number of states}

\item{trait.num}{The number of traits to simulate}

\item{br.rate}{clock Rate, currently can only be strict clock (a single rate)}
}
\description{
This function is to simulate discrete character data along the branches of a tree
}
\examples{
 # simulated tree
 phy <- ape::rtree(10)

 # simulate characters along the branches of the tree
 continuous_traits <- sim.morpho.completeprocess(phy, k=4, trait.num =12)

}
