% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.morpho.history.R
\name{sim.morpho.history}
\alias{sim.morpho.history}
\title{Simulate characters along branches in a tree}
\usage{
sim.morpho.history(
  tree = NULL,
  time.tree = NULL,
  ACRV = NULL,
  br.rates = NULL,
  ancestral = FALSE,
  k = 2,
  trait.num = 2,
  alpha.gamma = 1,
  ncats.gamma = 4
)
}
\arguments{
\item{tree}{Tree with branches that represent genetic distance associated with the character data.}

\item{time.tree}{Tree with branches that represent time associated with the character data.}

\item{ACRV}{Allow for among character rate variation. The default here is set to NULL. Can supply arguments "gamma" or "lognormal"}

\item{br.rates}{Clock rates per branch, currently can only be strict clock (a single rate)}

\item{ancestral}{Return the states at all ancestral nodes. Default set to false}

\item{k}{Number of trait states.}

\item{trait.num}{The number of traits to simulate}
}
\value{
An object of class morpho.
}
\description{
This function simulates discrete character data along the branches of a phylogentic tree. It can be used with
either a time tree or a tree with branch lengths in genetic distance. If using a time tree
branch rates can be specified, either as one values for all branches or as a vector with
different rates per branch. If no branch rates are specified a default of 1 is applied to
all branches.
}
\examples{
 # simulated tree
 phy <- ape::rtree(10)

 # simulate characters along the branches of the tree
 transition_history <-  sim.morpho.history(tree = phy, k = 3, trait.num = 4, ancestral = TRUE)

}
