% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.morpho.R
\name{sim.morpho}
\alias{sim.morpho}
\title{This function is to simulate discrete character data at the tips and internal nodes}
\usage{
sim.morpho(
  tree = NULL,
  time.tree = NULL,
  br.rates = NULL,
  k = 2,
  trait.num = 2,
  ancestral = FALSE
)
}
\arguments{
\item{tree}{Tree with branches that represent genetic distance associated with the character data.}

\item{time.tree}{Tree with branches that represent time associated with the character data.}

\item{k}{Number of states}

\item{trait.num}{The number of traits to simulate}

\item{ancestral}{T/F Whether or not to output information from internal nodes}

\item{data}{Trait data associated with the tips of the tree}

\item{br.rate}{clock Rate, currently can only be strict clock (a single rate)}
}
\description{
This function is to simulate discrete character data at the tips and internal nodes
}
