% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.r
\name{write.recon.matrix}
\alias{write.recon.matrix}
\title{Write reconstructed Matrix to file}
\usage{
write.recon.matrix(data, file = NULL, keep_matrix = F)
}
\arguments{
\item{data}{Morpho object}

\item{file}{File name}

\item{keep_matrix}{Logical. If TRUE, returns a matrix showing the naming transformations
between `morphosim` and `fossilsim` of Sampled ancestors.}
}
\description{
Write the reconstructed matrix to a nexus file
}
