% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.r
\name{write.recon.tree}
\alias{write.recon.tree}
\title{Write reconstructed tree to file}
\usage{
write.recon.tree(data, file)
}
\arguments{
\item{data}{Morpho object}

\item{file}{File name}
}
\description{
Write the reconstructed tree to Newick string
}
