% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.r
\name{write.tsv}
\alias{write.tsv}
\title{Write the taxa ages}
\usage{
write.tsv(data, file, uncertainty = 0)
}
\arguments{
\item{data}{Morpho object}

\item{file}{File name}

\item{uncertainty}{Numeric. Adds uncertainty to fossil ages in the morpho object.
The ages in the object are point estimates by default; setting `uncertainty`
will create an age range of ± this value (in millions of years).}
}
\description{
Writes the ages of the specimens in the true tree to a file. The tsv format used
here is directly compatible with RevBayes
}
