% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morpho.R
\name{morpho}
\alias{morpho}
\alias{as.morpho}
\alias{is.morpho}
\title{Morpho object}
\usage{
morpho(
  data = NULL,
  tree = NULL,
  time.tree = NULL,
  model = NULL,
  node.seq = NULL,
  transition_history = NULL,
  root.states = NULL,
  fossil_sequence = NULL,
  fossil = NULL,
  ACRV_rate = NULL,
  gamma_rates = NULL
)

as.morpho(
  data,
  tree,
  time.tree = NULL,
  model = NULL,
  node.seq = NULL,
  transition_history = NULL,
  root.states = NULL,
  fossil_sequence = NULL,
  fossil = NULL,
  ACRV_rate = NULL,
  gamma_rates = NULL
)

is.morpho(data)
}
\arguments{
\item{data}{Morphological data associated with the tips of the tree}

\item{tree}{Tree with branches that represent distances associated with the character data.}

\item{time.tree}{Tree with branches that represent time associated with the character data.}

\item{model}{Morphological model. (e.g. Mk, Mk+V, Mk+G)}

\item{node.seq}{Morphological data associated with the internal nodes of the tree}

\item{transition_history}{The constant character transitions along the branches}

\item{root.states}{A vector supplying the root state for each character}

\item{fossil_sequence}{Morphological data associated with the fossils on the tree}

\item{fossil}{Fossil object used to simulate data}

\item{ACRV_rate}{The rate category each trait was simulated under}

\item{gamma_rates}{The gamma rate for each rate category}
}
\description{
Create a morpho object.
}
