% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morpho.R
\name{morpho}
\alias{morpho}
\alias{as.morpho}
\alias{is.morpho}
\title{Mopho object}
\usage{
morpho(
  data = NULL,
  tree = NULL,
  time.tree = NULL,
  model = NULL,
  node.seq = NULL,
  continuous_traits = NULL,
  root.states = NULL
)

as.morpho(
  data,
  tree,
  time.tree = NULL,
  model = NULL,
  node.seq = NULL,
  continuous_traits = NULL,
  root.states = NULL
)

is.morpho(data)
}
\arguments{
\item{data}{Trait data associated with the tips of the tree}

\item{tree}{Tree with branches that represent genetic distance associated with the character data.}

\item{time.tree}{Tree with branches that represent time associated with the character data.}

\item{node.seq}{Trait data associated with the internal nodes of the tree}
}
\description{
Create a morpho.
}
