% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.morpho.grid.R
\name{plot.morpho.grid}
\alias{plot.morpho.grid}
\title{This function plots your morpho object to a character matrix}
\usage{
\method{plot}{morpho.grid}(
  morpho,
  xlab = "",
  ylab = "",
  name = "",
  col = c("white", "gray", "lightblue", "pink", "yellow", "green", "orange")
)
}
\arguments{
\item{morpho}{An object of class \code{"morpho"}.}

\item{xlab}{A string for the x-axis label (default: "Characters").}

\item{ylab}{A string for the y-axis label (default: "Taxa").}

\item{name}{A string for the legend title (default: "Character State").}

\item{col}{A character vector of colors. The length must match the number 
of unique character states in the data.}
}
\description{
This function creates a matrix from your morpho object highliting whic tips a which character state
}
\examples{
tree <- sim.bd.taxa(8,1,1,0.5)[[1]]
x <- sim.morpho.completeprocess(time.tree = tree, br.rates = 0.2, k = 2, num.traits = 5)
plot.morpho.grid(morpho = x)
}
