% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.morpho.R
\name{plot.morpho}
\alias{plot.morpho}
\title{Plot full evolutionary history}
\usage{
\method{plot}{morpho}(
  x = NULL,
  trait = NULL,
  timetree = FALSE,
  br.rates = NULL,
  fossil = FALSE,
  root.edge = FALSE,
  col = c("#fdfdfd", "lightgray", "lightblue", "pink", "yellow", "green", "orange"),
  col.timescale = "darkgrey",
  ...
)
}
\arguments{
\item{x}{A morpho object}

\item{trait}{What trait number you want to visualize}

\item{timetree}{TRUE or FALSE Indicate whether you want to plot a time tree or not. default FALSE, uses distance tree if FALSE}

\item{br.rates}{Required if you provide a time tree}

\item{fossil}{Do you want to plot the fossil along the tree. Default set to FALSE}

\item{root.edge}{If TRUE plot the root edge. Default = FALSE}

\item{col}{A vector of colors that should be the same length or longer than the number of different character states (k). if not specified, the traits from 0 to 6 can be differentiated}

\item{col.timescale}{a single color for the timescale, "darkgrey" by standard}

\item{...}{other arguments to be passed to methods, such as graphical parameters}
}
\description{
#' @description
This function creates a plot showing continuous evolution of discrete traits
}
\examples{
# Non-time tree
phy <- ape::rtree(10)

# simulate characters along the branches of the tree
transition_history <-  sim.morpho(tree = phy,
                                  k = c(2,3,4),
                                   trait.num = 20,
                                  ancestral = TRUE,
                                  partition = c(10,5,5),
                                 ACRV = "gamma",
                                 variable = TRUE,
                                  ncats.gamma = 4)

plot(x = transition_history, trait = 1, timetree = FALSE)

}
