% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.morpho.R
\name{plot.morpho}
\alias{plot.morpho}
\title{Plot full evolutionary history}
\usage{
\method{plot}{morpho}(
  x = NULL,
  trait = NULL,
  timetree = FALSE,
  br.rates = NULL,
  col = c("#fdfdfd", "lightgray", "lightblue", "pink", "yellow", "green", "orange"),
  col.timescale = "darkgrey",
  ...
)
}
\arguments{
\item{x}{A morpho object}

\item{trait}{What trait number you want to visualize}

\item{timetree}{TRUE or FALSE Indicate whether you want to plot a time tree or not. default FALSE, uses distance tree if FALSE}

\item{br.rates}{Required if you provide a time tree}

\item{col}{A vector of colors that should be the same length or longer than the number of different character states (k). if not specified, the traits from 0 to 6 can be differentiated}

\item{col.timescale}{a single color for the timescale, "darkgrey" by standard}

\item{...}{other arguments to be passed to methods, such as graphical parameters}
}
\description{
#' @description
This function creates a plot showing continuous evolution of discrete traits
}
\examples{
#set.seed(123)
# Non-time tree
phy <- ape::rtree(10)

# simulate characters along the branches of the tree
continuous_traits <- sim.morpho.completeprocess(phy, k=4, trait.num =12)

plot(x = continuous_traits, trait = 1, timetree = FALSE)

# time tree
tree <- TreeSim::sim.bd.taxa(8,1,1,0.5)[[1]]
continuous_traits <- sim.morpho.completeprocess(time.tree = tree, br.rates = 0.2,
k = 3, trait.num = 5)
plot(x = continuous_traits, trait = 1, timetree = TRUE, br.rates = 0.2,
col = c("#fdfdfd", "lightgray", "lightblue"), col.timescale = "black")
}
