% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.morpho.R
\name{plot.morpho}
\alias{plot.morpho}
\title{This function plots your morpho object to the tree}
\usage{
\method{plot}{morpho}(
  x,
  tree,
  show.tree = TRUE,
  max.age = NULL,
  show.axis = TRUE,
  rho = 1,
  l = 1,
  br.rates = 0.1,
  root.edge = TRUE,
  hide.edge = FALSE,
  edge.width = 1,
  show.tip.label = FALSE,
  align.tip.label = FALSE,
  extant.col = 1,
  taxa.palette = "viridis",
  col.axis = "gray35",
  cex = 1.2,
  pch = NULL,
  ...
)
}
\arguments{
\item{x}{an object of class "morpho"}

\item{tree}{an object of class "phy}

\item{show.tree}{Logical; whether to display the phylogenetic tree (default: TRUE).}

\item{max.age}{Numeric; maximum age for the time axis (default: NULL, inferred from tree).}

\item{br.rates}{Numeric; scaling factor for branch rates (default: 0.1).}

\item{edge.width}{Numeric; width of tree edges (default: 1).}

\item{show.tip.label}{Logical; whether to display tip labels (default: FALSE).}

\item{align.tip.label}{Logical; whether to align tip labels (default: FALSE).}

\item{extant.col}{Character; color for extant taxa points (default: "1").}

\item{taxa.palette}{Character; color palette for taxa (default: "viridis").}
}
\description{
this function plots a tree higlighting the evolutionary morphological change in regards to branch and time.
}
\examples{
tree <- sim.bd.taxa(8,1,1,0.5)[[1]]
x <- sim.morpho.completeprocess(time.tree = tree, br.rates = 0.2, k = 2, num.traits = 5)
plot.morpho(x = x, tree = x$tree)
}
