% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.morpho.R
\name{plot.morpho}
\alias{plot.morpho}
\title{Plot full evolutionary history}
\usage{
\method{plot}{morpho}(
  x = NULL,
  trait = NULL,
  timetree = FALSE,
  fossil = FALSE,
  reconstructed = FALSE,
  root.edge = FALSE,
  edges = 1,
  label.offset = 0.05,
  e.cex = 0.5,
  f.cex = 1,
  col = c("#fdfdfd", "lightgray", "lightblue", "pink", "yellow", "green", "orange"),
  col.timescale = "darkgrey",
  ...
)
}
\arguments{
\item{x}{A morpho object}

\item{trait}{The trait number to plot.}

\item{timetree}{TRUE or FALSE. Indicate whether you want to plot a time
tree or not. Default = FALSE (uses distance tree if FALSE).}

\item{fossil}{Plot the fossil along the tree. Default = FALSE.}

\item{reconstructed}{Plot the reconstructed tree. Default = FALSE.}

\item{root.edge}{If TRUE plot the root edge. Default = FALSE.}

\item{edges}{Width of the branch lengths.}

\item{label.offset}{Distance of tip label to tree tips.}

\item{e.cex}{Size of extant taxa.}

\item{f.cex}{Size of fossils.}

\item{col}{A vector of colors that should be the same length or longer than
the number of different character states (k). If not specified, the traits
from 0 to 6 can be differentiated.}

\item{col.timescale}{A single color for the timescale. Default = "darkgrey".}

\item{...}{Other arguments to be passed to methods, such as graphical
parameters.}
}
\description{
This function creates a plot showing continuous evolution of discrete traits.
}
\examples{
# simulate a phylogenetic tree
phy <- ape::rtree(10)

# simulate characters along the branches of the tree
morpho_data <- sim.morpho(
  tree = phy,
  k = c(2, 3, 4),
  trait.num = 20,
  ancestral = TRUE,
  partition = c(10, 5, 5),
  ACRV = "gamma",
  variable = TRUE,
  ACRV.ncats = 4,
  define.Q = NULL
)

plot(morpho_data, trait = 4, timetree = FALSE, fossil = FALSE,
     root.edge = FALSE, reconstructed = FALSE)

}
