% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recontruct.tree.R
\name{reconstruct.tree}
\alias{reconstruct.tree}
\title{#' @description
This function is used to colour the branches in the plotting
function when plotting the reconstructed tree.}
\usage{
reconstruct.tree(data)
}
\description{
#' @description
This function is used to colour the branches in the plotting
function when plotting the reconstructed tree.
}
