% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recontruct.tree.R
\name{reconstruct.tree}
\alias{reconstruct.tree}
\title{#' @description
This function is used to color the branches in the plotting
function when plotting the reconstructed tree.}
\usage{
reconstruct.tree(data)
}
\arguments{
\item{data}{Morpho object which contains fossil data}
}
\description{
#' @description
This function is used to color the branches in the plotting
function when plotting the reconstructed tree.
}
