% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.morpho.R
\name{sim.morpho}
\alias{sim.morpho}
\title{Simulate characters along branches in a tree}
\usage{
sim.morpho(
  tree = NULL,
  time.tree = NULL,
  ACRV = NULL,
  br.rates = NULL,
  variable = FALSE,
  ancestral = FALSE,
  k = 2,
  partition = NULL,
  trait.num = NULL,
  fossil = NULL,
  alpha.gamma = 1,
  ACRV.ncats = 4,
  define.gamma.rates = NULL,
  define.Q = NULL
)
}
\arguments{
\item{tree}{A phylogenetic tree (class "phylo") with branches representing genetic distance.}

\item{time.tree}{A phylogenetic tree (class "phylo") with branches representing time.}

\item{ACRV}{Among character rate variation using either `gamma`, `lgn`, or `user`. Default is `NULL`.}

\item{br.rates}{Clock rates per branch. Can be a single value (strict clock) or a vector of rates.}

\item{variable}{If `TRUE`, simulate only varying characters. Default is `FALSE`.}

\item{ancestral}{If `TRUE`, return the states at all ancestral nodes. Default is `FALSE`.}

\item{k}{Number of trait states (integer ≥ 2). Can be a vector if using partitions.}

\item{partition}{Specify the number of traits per partition}

\item{trait.num}{The total number of traits to simulate (integer > 0).}

\item{fossil}{Fossil object (from `FossilSim`) to simulate morphological characters.}

\item{alpha.gamma}{Shape parameter α for the gamma distribution.}

\item{ACRV.ncats}{Number of rate categories for among character rate variation.}

\item{define.gamma.rates}{Vector of gamma rate categories for the simulation.}

\item{define.Q}{Q matrix for simulation. Must be square and rows must sum to zero.}

\item{meanlog}{mean of the distribution on the log scale.}

\item{sdlog}{standard deviation of the distribution on the log scale.}
}
\value{
An object of class morpho.
}
\description{
This function simulates discrete character data along the branches of a phylogentic tree. It can be used with
either a time tree or a tree with branch lengths in evolutionary distance. If using a time tree
branch rates can be specified, either as one values for all branches or as a vector with
different rates per branch. If no branch rates are specified a default of 0.1 is applied to
all branches.
}
\examples{
# simulated tree
phy <- ape::rtree(10)

# simulate characters along the branches of the tree
transition_history <-  sim.morpho(tree = phy,
                                         k = c(2,3,4),
                                         trait.num = 20,
                                         ancestral = TRUE,
                                         partition = c(10,5,5),
                                         ACRV = "gamma",
                                         variable = TRUE,
                                         ACRV.ncats = 4)


# To simulate ordered characters:
# First define a Q-matrix. The following is for ordered characters where transitions can only occur
# between states 0 and 1 and 1 and 2

ord_Q <- matrix(c(
-0.5, 0.5, 0.0,
0.3333333, -0.6666667, 0.3333333,
0.0, 0.5, -0.5
), nrow = 3, byrow = TRUE)

# This Q matrix can be then used to simualte character data.

transition_history_2 <-  sim.morpho(tree = phy,
                                    k = 3,
                                    trait.num = 20,
                                    ancestral = TRUE,
                                    ACRV = "gamma",
                                    variable = TRUE,
                                    ACRV.ncats = 4,
                                  define.Q = ord_Q)
}
