% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.morpho.R
\name{sim.morpho}
\alias{sim.morpho}
\title{Simulate characters at internal nodes in a tree}
\usage{
sim.morpho(
  tree = NULL,
  time.tree = NULL,
  br.rates = NULL,
  k = 2,
  trait.num = 2,
  ancestral = FALSE
)
}
\arguments{
\item{tree}{Tree with branches that represent genetic distance associated with the character data.}

\item{time.tree}{Tree with branches that represent time associated with the character data.}

\item{k}{Number of states}

\item{trait.num}{The number of traits to simulate}

\item{ancestral}{T/F Whether or not to output information from internal nodes}

\item{data}{Trait data associated with the tips of the tree}

\item{br.rate}{clock Rate, currently can only be strict clock (a single rate)}
}
\value{
An object of class morpho.
}
\description{
This function simulates discrete character data at the internal nodes of a phylogentic tree. It can be used with
either a time tree or a tree with branch lengths in genetic distance. If using a time tree
branch rates can be specified, either as one values for all branches or as a vector with
different rates per branch. If no branch rates are specified a default of 1 is applied to
all branches.
}
\examples{
phy <- ape::rtree(10)
plot(phy)
simulated_morpho <- MorphoSim::sim.morpho(phy, k = 2, trait.num = 20)
simulated_morpho
}
