% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.r
\name{stats_morpho}
\alias{stats_morpho}
\title{Calculates statistics for a morpho object}
\usage{
stats_morpho(data)
}
\arguments{
\item{data}{Morpho object}
}
\description{
This function computes three key pieces of information:
1. The Consistency Index (CI) and Retention Index (RI) based on
the tip sequence data.
2. Convergent traits, identifying traits that have evolved independently
 multiple times.
3. Summary information about the size and structure of the tree.
}
\examples{
#' # simulate a phylogenetic tree
phy <- ape::rtree(10)

# simulate characters along the branches of the tree
morpho_data <- sim.morpho(tree = phy,
                          k = c(2,3,4),
                          trait.num = 20,
                          ancestral = TRUE,
                          partition = c(10,5,5),
                          ACRV = "gamma",
                          variable = TRUE,
                          ACRV.ncats = 4)

summary <- stats_morpho(data = morpho_data)
}
